[Alexandra, Bruno] Refatora o server
