+++ src/App.css
@@ -47,9 +47,6 @@ ul li {
   width: 230px;
   height: 50px;
   padding: 12px;
-}
-
-.btn.btn-secondary.btn-raised p {
   color: white;
   font-size: 16px;
   font-weight: 600;
