+++ src/components/Login/Login.test.js
@@ -8,13 +8,13 @@ configure({ adapter: new Adapter() });
 describe('<Login />', () => {
   it('should render the login button', () => {
     let wrapper = shallow(<Login loginAction={() => {}} />);
-    expect(wrapper.find('a').text()).toEqual('Login');
+    expect(wrapper.find('button').text()).toEqual('Login');
   });
 
   it('Should click the login on the button', () => {
     const mockAction = jest.fn();
     let wrapper = shallow(<Login loginAction={mockAction} />);
-    wrapper.find('a').simulate('click');
+    wrapper.find('button').simulate('click');
 
     expect(mockAction.mock.calls.length).toEqual(1);
   });
