+++ src/server/clientFactory.js
@@ -1,0 +1,12 @@
+const { OAuth2Client } = require('google-auth-library');
+
+const {
+  GOOGLE_CLIENT_ID,
+  GOOGLE_CLIENT_SECRET,
+  GOOGLE_REDIRECT_URL,
+} = process.env;
+
+const googleOAuth2Client = () =>
+  new OAuth2Client(GOOGLE_CLIENT_ID, GOOGLE_CLIENT_SECRET, GOOGLE_REDIRECT_URL);
+
+module.exports = { googleOAuth2Client };
